+++ src/main/java/site/scalarstudios/scalarpower/item/ScalarPowerItems.java
@@ -12,7 +12,9 @@ public class ScalarPowerItems {
     // Dusts
     public static final DeferredItem<Item> COAL_DUST = ITEMS.registerSimpleItem("coal_dust");
     public static final DeferredItem<Item> COPPER_DUST = ITEMS.registerSimpleItem("copper_dust");
+    public static final DeferredItem<Item> COBALT_DUST = ITEMS.registerSimpleItem("cobalt_dust");
     public static final DeferredItem<Item> DIAMOND_DUST = ITEMS.registerSimpleItem("diamond_dust");
+    public static final DeferredItem<Item> ENDER_ALLOY_DUST = ITEMS.registerSimpleItem("ender_alloy_dust");
     public static final DeferredItem<Item> EMERALD_DUST = ITEMS.registerSimpleItem("emerald_dust");
     public static final DeferredItem<Item> GOLD_DUST = ITEMS.registerSimpleItem("gold_dust");
     public static final DeferredItem<Item> IRON_DUST = ITEMS.registerSimpleItem("iron_dust");
@@ -23,11 +25,14 @@ public class ScalarPowerItems {
     public static final DeferredItem<Item> SAWDUST = ITEMS.registerSimpleItem("sawdust");
 
     // Chunks
+    public static final DeferredItem<Item> COBALT_CHUNK = ITEMS.registerSimpleItem("cobalt_chunk");
     public static final DeferredItem<Item> COPPER_CHUNK = ITEMS.registerSimpleItem("copper_chunk");
     public static final DeferredItem<Item> IRON_CHUNK = ITEMS.registerSimpleItem("iron_chunk");
     public static final DeferredItem<Item> GOLD_CHUNK = ITEMS.registerSimpleItem("gold_chunk");
 
     // Ingots
+    public static final DeferredItem<Item> COBALT_INGOT = ITEMS.registerSimpleItem("cobalt_ingot");
+    public static final DeferredItem<Item> ENDER_ALLOY_INGOT = ITEMS.registerSimpleItem("ender_alloy_ingot");
     public static final DeferredItem<Item> REDIUM_INGOT = ITEMS.registerSimpleItem("redium_ingot");
     public static final DeferredItem<Item> STEEL_INGOT = ITEMS.registerSimpleItem("steel_ingot");